test/scenario/multi-backend/requirements.txt
fabric
python-swiftclient
